requirements.txt
absl-py==2.1.0
acme==1.21.0
alembic==1.14.1
apturl==0.5.2
astunparse==1.6.3
attrs==21.2.0
blinker==1.9.0
Brlapi==0.8.3
certbot==1.21.0
certifi==2020.6.20
chardet==4.0.0
chex==0.1.89
click==8.1.8
colorama==0.4.4
command-not-found==0.3
ConfigArgParse==1.5.3
configobj==5.0.6
contourpy==1.3.1
cryptography==3.4.8
cupshelpers==1.0
cycler==0.12.1
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
etils==1.12.2
Far==0.0.1
filelock==3.16.1
Flask==3.1.0
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
flax==0.10.4
fonttools==4.54.1
fsspec==2024.10.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.1.1
grpcio==1.71.0
h5py==3.13.0
httplib2==0.20.2
humanize==4.12.1
idna==3.3
importlib-metadata==4.6.4
importlib_resources==6.5.2
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
jeepney==0.7.1
Jinja2==3.1.4
josepy==1.10.0
jsonschema==3.2.0
keras==3.9.0
keyring==23.5.0
kiwisolver==1.4.7
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
libclang==18.1.1
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml_dtypes==0.5.1
more-itertools==8.10.0
mpmath==1.3.0
msgpack==1.1.0
namex==0.0.8
nest-asyncio==1.6.0
netifaces==0.11.0
networkx==3.4.2
numpy==2.1.1
nvidia-cublas-cu11==11.11.3.6
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu11==9.1.0.70
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu11==10.3.0.86
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu11==11.7.5.86
nvidia-cusparse-cu12==12.3.1.170
nvidia-nccl-cu11==2.21.5
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu11==11.8.86
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.0
olefile==0.46
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
opt_einsum==3.4.0
optax==0.2.4
optree==0.14.1
orbax-checkpoint==0.11.9
packaging==23.2
pandas==2.2.3
parsedatetime==2.6
pexpect==4.8.0
Pillow==9.0.1
protobuf==5.29.4
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
py-cpuinfo==9.0.0
pycairo==1.20.1
pycups==2.0.1
Pygments==2.19.1
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.9.0.post0
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==0.9.1
rich==13.9.4
scipy==1.15.2
seaborn==0.13.2
SecretStorage==3.3.1
simplejson==3.20.1
six==1.16.0
SQLAlchemy==2.0.38
ssh-import-id==5.11
sympy==1.13.1
systemd-python==234
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
tensorflow-io-gcs-filesystem==0.37.1
tensorflow_decision_forests==1.12.0
tensorflowjs==4.22.0
tensorstore==0.1.72
termcolor==2.5.0
texttable==1.6.4
tf_keras==2.19.0
toolz==1.0.0
torch==2.5.1+cu118
torchaudio==2.5.1+cu118
torchvision==0.20.1+cu118
tqdm==4.67.1
treescope==0.1.9
triton==3.1.0
typing_extensions==4.12.2
tzdata==2025.1
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
ultralytics==8.3.93
ultralytics-thop==2.0.14
unattended-upgrades==0.1
urllib3==1.26.5
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.1.3
wrapt==1.17.2
wurlitzer==3.1.1
xdg==5
xkit==0.0.0
ydf==0.11.0
zipp==1.0.0
zope.component==4.3.0
zope.event==4.4
zope.hookable==5.1.0
zope.interface==5.4.0
